+++ src/gameobjects/GameObject.ts
@@ -13,7 +13,8 @@ export default class GameObject {
     public realScale: number = scale,
     public data: any = null,
     public ownerSID: number = -1,
-    public health: number = -1
+    public health: number = -1,
+    public dmg: number = 0
   ) {}
 
   getData() {