+++ CHANGELOG.md
@@ -2,6 +2,54 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.1.0](https://github.com/Picoseconds/sanctuary/compare/v1.0.0...v1.1.0) (2020-08-16)
+
+
+### Features
+
+* add /weaponVariant functionality for other players ([8d3983a](https://github.com/Picoseconds/sanctuary/commit/8d3983a9817de6e8f9ef355f7f4a6a3ee9546dcd))
+* add CPS cap ([bf8db3f](https://github.com/Picoseconds/sanctuary/commit/bf8db3f18d82e4c504903f9cb600c06b3b4695c0))
+* add miner's helmet ([bfecc71](https://github.com/Picoseconds/sanctuary/commit/bfecc711cfb17f58699953ff4b366de13f80e4b7))
+* add speed changes for holding items ([2021197](https://github.com/Picoseconds/sanctuary/commit/2021197cb39e52ebdd623cef4d81679dc5aa880a))
+* add toggle for invincibility ([b4b609c](https://github.com/Picoseconds/sanctuary/commit/b4b609ceedf323597023dc643e8d3d90faaf4034))
+* added /invisible command ([d61f4d3](https://github.com/Picoseconds/sanctuary/commit/d61f4d3c5d36a39d9bec1ec672d0bda30da02d82))
+* added /tp command to teleport to another player ([7580849](https://github.com/Picoseconds/sanctuary/commit/75808494460fb0924ae287f7e876c2fef11981fa))
+* added /weaponVariant command ([403fec7](https://github.com/Picoseconds/sanctuary/commit/403fec7269b0434190bc1302cda90b81c849e32e))
+* added cacti ([4afbc3a](https://github.com/Picoseconds/sanctuary/commit/4afbc3a47d535a0f8910bd40cd0010938399cd90))
+* added IP limit due to current bot attack ([cb99b30](https://github.com/Picoseconds/sanctuary/commit/cb99b306b1d51a9a499d140e4822e980c9cc77e7))
+* added pit traps and spikes ([69d1f08](https://github.com/Picoseconds/sanctuary/commit/69d1f08f766cbdbbe4fa9ed81646055062836c6e))
+* added place limits ([5e6f49a](https://github.com/Picoseconds/sanctuary/commit/5e6f49a2509fd919e98314607b834fd27a9ab724))
+* added rudimentary API ([e90a532](https://github.com/Picoseconds/sanctuary/commit/e90a532ad4a0eff842aa4ba3ddcc46a9dfa2d534))
+* added samurai armor ([d3b61c3](https://github.com/Picoseconds/sanctuary/commit/d3b61c37ace10a439f130d1fe92314e30829d30e))
+* added speed multiplier command for admins ([586e293](https://github.com/Picoseconds/sanctuary/commit/586e293b1994a861b690d669a2f9650c872a5acb))
+* hide pit traps from enemies ([1a37e44](https://github.com/Picoseconds/sanctuary/commit/1a37e44b20b2fccc879416b4d5af02689326f713))
+* made admins invincible ([4c3d01a](https://github.com/Picoseconds/sanctuary/commit/4c3d01a61d3b1d392cac6a0ed9f2ade29731769b))
+* made invisible players not show up in leaderboard ([451354e](https://github.com/Picoseconds/sanctuary/commit/451354e684d59ec75ea32106f1bbcabfb134b1ce))
+* made spikes show damage to the owner ([16f9cd2](https://github.com/Picoseconds/sanctuary/commit/16f9cd2898ed7314a88b5ac25672ee73594c3e77))
+
+
+### Bug Fixes
+
+* fix collision testing ([f87378e](https://github.com/Picoseconds/sanctuary/commit/f87378e2205d8b69e1dc0f1718959a5cda4ceb98))
+* fix damage coefficient ([e77a85b](https://github.com/Picoseconds/sanctuary/commit/e77a85b3e4ff2c249225a68d5ee8d2dd4ab84218))
+* fix IP limit ([8b3af02](https://github.com/Picoseconds/sanctuary/commit/8b3af02430089dd9fa0f9e1f7731170bb5e98eb6))
+* fix items that don't collide like traps and pads ([89d0128](https://github.com/Picoseconds/sanctuary/commit/89d012877bfa4474af6f5eb4d164809a0dfdf3b6))
+* fix promoting leaving user with broken age ([15e7b99](https://github.com/Picoseconds/sanctuary/commit/15e7b99edfa3835d4f7981bc90bc33dd6933bc9c))
+* fix some things not resetting on death ([062aa57](https://github.com/Picoseconds/sanctuary/commit/062aa57f641bdcc7703fc19054ec0f0c958bacc1))
+* fix spikes and other gameobjects not damaging until the player moves ([f706528](https://github.com/Picoseconds/sanctuary/commit/f706528759eec928b53fabbc348998e652ab0952))
+* fix things turning into apples when placed ([025744d](https://github.com/Picoseconds/sanctuary/commit/025744d790e6f1e89949a8c2d7500cc78c3ae55b))
+* fixed API stringifying data ([ae7b958](https://github.com/Picoseconds/sanctuary/commit/ae7b9580074e8f31ea79d08bd8deae02863169d9))
+* fixed client removal ([89611e4](https://github.com/Picoseconds/sanctuary/commit/89611e49752068ac30dd4bc1c861f843ae4fcc31))
+* fixed invincible admins ([2ae3e85](https://github.com/Picoseconds/sanctuary/commit/2ae3e859ce48e2513ba5ce76e18dd63134bbad19))
+* fixed pit trap hitbox being too small ([e62567d](https://github.com/Picoseconds/sanctuary/commit/e62567df5c4aadec56a69c2020bd6b1930f1459d))
+* fixed your own spikes making you stuck ([23b04b5](https://github.com/Picoseconds/sanctuary/commit/23b04b568c04255d99e9386000485dedd6d6c327))
+* limited gameobject types in the desert and river ([57a8dad](https://github.com/Picoseconds/sanctuary/commit/57a8dad93bd2ba997553924013bfca9fbf91d908))
+* make gold give xp ([69ae80b](https://github.com/Picoseconds/sanctuary/commit/69ae80b74dd7c07587485388ad46a0935e7ccd51))
+* protected invincible admins from player attacks ([fe5a0b0](https://github.com/Picoseconds/sanctuary/commit/fe5a0b0b472a3821861103a8a3d7e0deaf202a13))
+* put missing semicolon in Physics.ts ([26a3e83](https://github.com/Picoseconds/sanctuary/commit/26a3e8334fba42b376832ecc57446197bb505570))
+* remove clients and their placed things after they leave ([23fbcf7](https://github.com/Picoseconds/sanctuary/commit/23fbcf7f081f02b913ed7230a545fe6582ed44b3))
+* **component:** remove trailing slashes from socket pathnames ([82c98f2](https://github.com/Picoseconds/sanctuary/commit/82c98f204771e4180ae4bf2d431ddc1bd4914815))
+
 ## 1.0.0 (2020-08-02)